# perf: optimize certain edge cases