~38% improvement.

## Before

```
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark           Time             CPU   Iterations        GHz cycle/byte cycles/mime instructions/byte instructions/cycle instructions/mime instructions/ns     mime/s    ns/mime      speed  time/byte  time/mime
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench     102512 ns       101567 ns         6955    3.19387    31.7451     357.844           164.808            5.19159          1.85778k         16.5813 8.67411M/s    112.041 97.7782M/s  10.2272ns  115.286ns

```

## After

```
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark           Time             CPU   Iterations        GHz cycle/byte cycles/mime instructions/byte instructions/cycle instructions/mime instructions/ns     mime/s    ns/mime      speed  time/byte  time/mime
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench      72347 ns        72346 ns         9677    3.27756    23.7897     268.168           128.236            5.39037          1.44553k         17.6672 12.1775M/s    81.8195  137.27M/s  7.28492ns  82.1186ns

```